Added Doors
Added Doors for locking/unlocking areas. Does not have a script but the animation player has the animations ready.

To be used with: https://github.com/tristenk94/DMCIVG/issues/38

May need to modify the animation sprites to get rid of the flooring since it draws the sprite over the player/enemies and looks a little funny. Will leave it for now